Add support for extracting from a stream
Add support for extracting from a stream. This benefits the case when decoding data directly from file as there is no longer the need to load the full contents of the file to an intermediary string.

This is a behaviour changing modification with API backwards compatibility.
The following behaviour has changed:
the content of the error string when a parse error is encountered may be different.

This involves the following internal changes:
Change MsgPackParser from class to namespace as state is now effectively held in the stream.
Modify parse_* to read data from the stream.
Change read_bytes() to a templated function to simplify zeroing result in case of stream error.

The following additional changes have been made:
Add test to verify that incomplete data is correctly detected and error'd on.